Refactor hooks

+++ src/flowplayer.tsx
@@ -4,18 +4,19 @@ import flowplayer, { Flowplayer } from "@flowplayer/player"
 const Flowplayer = React.forwardRef<HTMLDivElement, Flowplayer.Config>((opts, forwardedRef) => {
     const ref = forwardedRef === null ? useRef(null) : forwardedRef
 
-    const { src, ...rest } = opts
+    const { token, src, ...rest } = opts
     // Init Flowplayer on mount
     useEffect(() => {
         if (typeof ref === "function") return
         if (!ref) return
         if (!ref.current) return
-        const api = flowplayer(ref.current, {...opts})
+        if (!token) return
+        const api = flowplayer(ref.current, {token,...opts})
         return () => {
             api.destroy()
             if (ref.current) ref.current.innerHTML = ""
         }
-    }, [ref])
+    }, [token, ref])
 
     const useApi = () => ref && typeof ref !== "function" &&  ref.current ? flowplayer(ref.current) : null
